# docs/swagger.yaml
definitions:
...
  controllers.LoginInput:
    properties:
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterInput:
    properties:
      city:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
      username:
        type: string
    required:
    - city
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  controllers.UserCategoryInput:
    properties:
      category:
        items:
          type: string
        type: array
      userID:
        type: integer
    type: object
...
host: www.tripwise.my.id
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Documentation TripWise Capstone Project - Team 3
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TRIPWISE API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Handle user login by verifying username and password, and return
        a JWT token.
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - User
  /logout:
    get:
      consumes:
      - application/json
      description: Process the logout request and return a success message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Log out a user
      tags:
      - User
  /register:
    post:
      consumes:
      - application/json
      description: Handle user registration by validating input and creating a new
        user in the database.
      parameters:
      - description: Registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User registration
      tags:
      - User
...
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information for a specific user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get user details
      tags:
      - User
  /user/category:
    post:
      consumes:
      - application/json
      description: Assign categories to a user by updating their profile.
      parameters:
      - description: User categories
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.UserCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create or update user categories
      tags:
      - User
  /users/{id}:
    put:
      consumes:
      - multipart/form-data
      description: Update user information such as username, email, password, and
        categories.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Username
        in: formData
        name: username
        type: string
      - description: First Name
        in: formData
        name: first_name
        type: string
      - description: Last Name
        in: formData
        name: last_name
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: City
        in: formData
        name: city
        type: string
      - description: Password
        in: formData
        name: password
        type: string
      - description: Role
        in: formData
        name: role
        type: string
      - description: Phone Number
        in: formData
        name: phone_number
        type: string
      - description: Gender
        in: formData
        name: gender
        type: string
      - description: Profile Image
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Edit user profile
      tags:
      - User
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"